Revert "Refactor search logic"
