+++ src/components/config.js
@@ -292,130 +292,6 @@ export const fetchDislikeUsersData = async ownerId => {
   }
 };
 
-export const setUserComment = async (ownerId, userId, comment) => {
-  try {
-    await set(ref2(database, `multiData/comments/${ownerId}/${userId}`), comment);
-  } catch (error) {
-    console.error('Error setting comment:', error);
-  }
-};
-
-export const fetchUserComment = async (ownerId, userId) => {
-  try {
-    const snap = await get(ref2(database, `multiData/comments/${ownerId}/${userId}`));
-    return snap.exists() ? snap.val() : '';
-  } catch (error) {
-    console.error('Error fetching comment:', error);
-    return '';
-  }
-};
-
-const addUserFromUsers = async (userId, users) => {
-  const snap = await get(ref2(database, `users/${userId}`));
-  if (snap.exists()) {
-    users[userId] = { userId, ...snap.val() };
-  }
-};
-
-const searchBySearchIdUsers = async (modifiedSearchValue, uniqueUserIds, users) => {
-  const ukSmPrefix = encodeKey('УК СМ ');
-  const hasUkSm = modifiedSearchValue.toLowerCase().startsWith(ukSmPrefix.toLowerCase());
-  const searchPromises = keysToCheck.flatMap(prefix => {
-    const baseKey = `${prefix}_${modifiedSearchValue.toLowerCase()}`;
-    const searchKeys = [baseKey];
-    if (hasUkSm) {
-      const withoutPrefix = modifiedSearchValue.slice(ukSmPrefix.length).toLowerCase();
-      searchKeys.push(`${prefix}_${withoutPrefix}`);
-    } else {
-      searchKeys.push(`${prefix}_${ukSmPrefix.toLowerCase()}${modifiedSearchValue.toLowerCase()}`);
-    }
-    if (modifiedSearchValue.startsWith('0')) searchKeys.push(`${prefix}_38${modifiedSearchValue.toLowerCase()}`);
-    if (modifiedSearchValue.startsWith('+')) searchKeys.push(`${prefix}_${modifiedSearchValue.slice(1).toLowerCase()}`);
-    return searchKeys.map(async sk => {
-      const snap = await get(query(ref2(database, 'searchId'), orderByKey(), startAt(sk), endAt(`${sk}\uf8ff`)));
-      if (snap.exists()) {
-        for (const [, val] of Object.entries(snap.val())) {
-          const ids = Array.isArray(val) ? val : [val];
-          for (const id of ids) {
-            if (!uniqueUserIds.has(id)) {
-              uniqueUserIds.add(id);
-              await addUserFromUsers(id, users);
-            }
-          }
-        }
-      }
-    });
-  });
-  await Promise.all(searchPromises);
-};
-
-const searchByPrefixesUsers = async (searchValue, uniqueUserIds, users) => {
-  for (const prefix of keysToCheck) {
-    let formatted = searchValue.trim().toLowerCase();
-    if (prefix === 'name' || prefix === 'surname') {
-      formatted = searchValue.trim().charAt(0).toUpperCase() + searchValue.trim().slice(1).toLowerCase();
-    }
-    const q = query(ref2(database, 'users'), orderByChild(prefix), startAt(formatted), endAt(`${formatted}\uf8ff`));
-    try {
-      const snap = await get(q);
-      if (snap.exists()) {
-        snap.forEach(userSnap => {
-          const userId = userSnap.key;
-          const userData = userSnap.val();
-          let fieldValue = userData[prefix];
-          if (typeof fieldValue === 'string') fieldValue = fieldValue.trim();
-          else return;
-          if (fieldValue && fieldValue.toLowerCase().includes(formatted.toLowerCase()) && !uniqueUserIds.has(userId)) {
-            uniqueUserIds.add(userId);
-            users[userId] = { userId, ...userData };
-          }
-        });
-      }
-    } catch {}
-  }
-};
-
-const searchUserByPartialUserIdUsers = async (userId, users) => {
-  try {
-    const refToCollection = ref2(database, 'users');
-    const partialQuery = query(refToCollection, orderByKey(), startAt(userId), endAt(userId + '\uf8ff'));
-    const snapshot = await get(partialQuery);
-    if (snapshot.exists()) {
-      snapshot.forEach(userSnapshot => {
-        const currentUserId = userSnapshot.key;
-        if (currentUserId.includes(userId)) {
-          users[currentUserId] = { userId: currentUserId, ...userSnapshot.val() };
-        }
-      });
-      if (Object.keys(users).length > 0) return users;
-    }
-    return null;
-  } catch (error) {
-    console.error('Error fetching data by partial userId:', error);
-    return null;
-  }
-};
-
-export const searchUsersOnly = async searchedValue => {
-  const { searchValue, modifiedSearchValue } = makeSearchKeyValue(searchedValue);
-  const users = {};
-  const uniqueUserIds = new Set();
-  try {
-    await searchBySearchIdUsers(modifiedSearchValue, uniqueUserIds, users);
-    await searchByPrefixesUsers(searchValue, uniqueUserIds, users);
-    await searchUserByPartialUserIdUsers(searchValue, users);
-    if (Object.keys(users).length === 1) {
-      const id = Object.keys(users)[0];
-      return users[id];
-    }
-    if (Object.keys(users).length > 1) return users;
-    return {};
-  } catch (error) {
-    console.error('Error searching users only:', error);
-    return null;
-  }
-};
-
 export const makeNewUser = async searchedValue => {
   const db = getDatabase();
   const newUsersRef = ref2(db, 'newUsers');
